+++ frontend/src/createStream/StreamingPage.jsx
@@ -3,7 +3,7 @@ import { Box, Container } from "@material-ui/core";
 import PropTypes from "prop-types";
 import StreamingTable from "./StreamingTable";
 import DynamicBreadcrumb from "../general/DynamicBreadcrumb";
-import StreamList from "../streamlist/StreamList";
+import StreamList from "../streamlist/StreamListPage";
 
 export default class StreamingPage extends React.Component {
   constructor(props) {
